requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1
scikit-learn==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyMySQL==1.1.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
opencv-python==4.8.1.78
faiss-cpu==1.7.4
pydantic==2.5.0
aiofiles==23.2.1
git+https://github.com/openai/CLIP.git